# apps/backends/socialweb-profiles-service/docs/swagger.json
{"schemes":["http","https"],"swagger":"2.0","info":{"description":"This is an API of profiles-service","title":"AXA-socialweb-service API (profiles)","contact":{},"version":"1.0"},"basePath":"/api/v1","paths":{"/profile/chat-sync":{"put":{"security":[{"CookieAuth":[]}],"description":"Synchronization with chat service","consumes":["application/json"],"produces":["application/json"],"tags":["profile"],"summary":"Synchronization with chat service","responses":{"200":{"description":"OK","schema":{"type":"string"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/types.FailureErrorResponse"}}}}},"/profile/me":{"get":{"security":[{"CookieAuth":[]}],"description":"Profile","consumes":["application/json"],"produces":["application/json"],"tags":["profile"],"summary":"Profile","responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/types.Profile"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/types.FailureErrorResponse"}}}},"put":{"security":[{"CookieAuth":[]}],"description":"Profile update","consumes":["application/json"],"produces":["application/json"],"tags":["profile"],"summary":"Profile update","parameters":[{"description":"info for updating","name":"request","in":"body","required":true,"schema":{"$ref":"#/definitions/types.ProfileUpdateRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/types.Profile"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/types.FailureErrorResponse"}}}},"patch":{"security":[{"CookieAuth":[]}],"description":"Profile update","consumes":["application/json"],"produces":["application/json"],"tags":["profile"],"summary":"Profile update","parameters":[{"description":"info for updating","name":"request","in":"body","required":true,"schema":{"$ref":"#/definitions/types.ProfileUpdateRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/types.Profile"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/types.FailureErrorResponse"}}}}},"/profile/me/set-avatar":{"put":{"security":[{"CookieAuth":[]}],"description":"Set profile avatar","consumes":["multipart/form-data"],"produces":["application/json"],"tags":["profile"],"summary":"Set profile avatar","parameters":[{"type":"file","description":"this is a test file","name":"file","in":"formData","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/types.FileUploadResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/types.FailureErrorResponse"}}}}},"/profile/{username}/public":{"get":{"description":"Get public profile by username","consumes":["application/json"],"produces":["application/json"],"tags":["profile"],"summary":"Get public profile","parameters":[{"type":"string","description":"Username","name":"username","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/types.PublicProfile"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/types.FailureErrorResponse"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/types.FailureErrorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/types.FailureErrorResponse"}}}}}},"definitions":{"types.FailureErrorResponse":{"type":"object","properties":{"error":{"type":"string"},"message":{"type":"string"},"status":{"type":"string"}}},"types.FileUploadResponse":{"type":"object","properties":{"url":{"type":"string"}}},"types.Post":{"type":"object","properties":{"content":{"type":"string"},"files":{"type":"array","items":{"type":"string"}},"id":{"type":"string"},"media_url":{"type":"string"},"payment":{"type":"boolean"},"subscribers":{"type":"array","items":{"type":"string"}},"tags":{"type":"array","items":{"type":"string"}},"title":{"type":"string"},"type":{"type":"string"},"user_id":{"type":"string"}}},"types.Profile":{"type":"object","properties":{"avatar_url":{"type":"string"},"cover_url":{"type":"string"},"description":{"type":"string"},"donation_url":{"type":"string"},"email":{"type":"string"},"email_confirmed":{"type":"boolean"},"id":{"type":"string"},"notification_settings":{"$ref":"#/definitions/types.ProfileNotificationSettings"},"permits":{"type":"array","items":{"type":"string"}},"roles":{"type":"array","items":{"type":"string"}},"username":{"type":"string"}}},"types.ProfileNotificationSettings":{"description":"Настройки уведоплений для пользователя например \"Включить звук уведомления\", \"Когда кто-то подписался на вас\" и т.д.","type":"object","properties":{"desktop_notifications":{"description":"Показывать уведомления на рабочем столе","type":"boolean"},"enable_notifications":{"description":"Включить/Выключить нотификации","type":"boolean"},"notification_sound":{"description":"Включить звук уведомления","type":"boolean"},"suspicious_attempts_to_logon":{"description":"Письма о подозрительных попытках войти в вашу учётную запись","type":"boolean"},"when_idea_update_email":{"description":"Когда есть обновление (Идеи, на которые вы подписаны) email","type":"boolean"},"when_idea_update_ws":{"description":"Когда есть обновление (Идеи, на которые вы подписаны) web","type":"boolean"},"when_script_update_email":{"description":"Когда есть обновление (Скрипты, которые вы добавили в Избранное или оценили) email","type":"boolean"},"when_script_update_ws":{"description":"Когда есть обновление (Скрипты, которые вы добавили в Избранное или оценили) web","type":"boolean"},"when_someone_comments_you_idea_email":{"description":"Когда кто-то комментирует вашу идею (Ваши идеи) email","type":"boolean"},"when_someone_comments_you_idea_ws":{"description":"Когда кто-то комментирует вашу идею (Ваши идеи) web","type":"boolean"},"when_someone_follows_you_email":{"description":"Когда кто-то подписался на вас (Ваш профиль) email","type":"boolean"},"when_someone_follows_you_ws":{"description":"Когда кто-то подписался на вас (Ваш профиль) web","type":"boolean"},"when_someone_likes_you_idea_email":{"description":"Когда кому-то понравилась ваша идея (Ваши идеи) email","type":"boolean"},"when_someone_likes_you_idea_ws":{"description":"Когда кому-то понравилась ваша идея (Ваши идеи) web","type":"boolean"},"when_someone_mentions_you_in_idea_comment_ws":{"description":"Когда кто-то упоминает вас в комментарии к мнению (Мнения) web","type":"boolean"},"when_someone_mentions_you_in_idea_comments_ws":{"description":"Когда кто-то упоминает вас в комментариях к идее (Ваш профиль) web","type":"boolean"},"when_someone_mentions_you_in_opinion_email":{"description":"Когда кто-то упоминает вас в мнении (Мнения) email","type":"boolean"},"when_someone_mentions_you_in_opinion_ws":{"description":"Когда кто-то упоминает вас в мнении (Мнения) web","type":"boolean"},"when_someone_mentions_you_while_you_offline_email":{"description":"Когда вас упомянули в чате, пока вы были не в сети (Ваш профиль) email","type":"boolean"},"when_someone_mentions_you_while_you_offline_ws":{"description":"Когда вас упомянули в чате, пока вы были не в сети (Ваш профиль) web","type":"boolean"},"when_they_publish_new_idea_email":{"description":"Когда они опубликуют новую идею (Авторы, на которых вы подписаны) email","type":"boolean"},"when_they_publish_new_idea_ws":{"description":"Когда они опубликуют новую идею (Авторы, на которых вы подписаны) web","type":"boolean"},"when_they_publish_new_opinion_email":{"description":"Когда они публикуют новое мнение (Авторы, на которых вы подписаны) email","type":"boolean"},"when_they_publish_new_opinion_ws":{"description":"Когда они публикуют новое мнение (Авторы, на которых вы подписаны) web","type":"boolean"}}},"types.ProfileUpdateRequest":{"type":"object","properties":{"description":{"type":"string"},"donation_url":{"type":"string"},"notification_settings":{"$ref":"#/definitions/types.ProfileNotificationSettings"},"username":{"type":"string"}}},"types.PublicProfile":{"type":"object","properties":{"avatar_url":{"type":"string"},"bio":{"type":"object"},"followers":{"type":"array","items":{"$ref":"#/definitions/types.UserInfo"}},"followings":{"type":"array","items":{"$ref":"#/definitions/types.UserInfo"}},"id":{"type":"string"},"member_since":{"type":"string"},"posts":{"type":"array","items":{"$ref":"#/definitions/types.Post"}},"username":{"type":"string"}}},"types.UserInfo":{"type":"object","properties":{"avatar":{"type":"string"},"cover":{"type":"string"},"created":{"type":"string"},"description":{"type":"string"},"donation":{"type":"string"},"email":{"type":"string"},"email_confirmed":{"type":"boolean"},"id":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"updated":{"type":"string"},"username":{"type":"string"}}}},"securityDefinitions":{"CookieAuth":{"description":"Сессионная аутентификация через куку sessionid","type":"apiKey","name":"sessionid","in":"cookie"}}}